Added new database project

+++ telegram-bot-api/src/main/java/de/simonscholz/bot/telegram/api/Location.java
@@ -7,6 +7,6 @@ import lombok.Data;
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Location {
-	private float longitude;
-	private float latitude;
+	private double longitude;
+	private double latitude;
 }
