+++ telegram-bot-springboot/src/main/java/de/simonscholz/bot/telegram/location/OpenStreetMapApi.java
@@ -8,5 +8,5 @@ public interface OpenStreetMapApi {
 	static final String BASE_URL = "http://nominatim.openstreetmap.org/";
 
 	@GET("reverse?format=jsonv2")
-	Single<OSMLocation> getLocationData(@Query("lat") float latitude, @Query("lon") float longitude);
+	Single<OSMLocation> getLocationData(@Query("lat") double latitude, @Query("lon") double longitude);
 }
